event pos use Vec2i

+++ include/flame/universe/components/event_receiver.h
@@ -31,8 +31,8 @@ namespace flame
 
 		FLAME_UNIVERSE_EXPORTS void* add_focus_listener(void (*listener)(void* c, FocusType type), const Mail<>& capture);
 		FLAME_UNIVERSE_EXPORTS void* add_key_listener(void (*listener)(void* c, KeyState action, int value), const Mail<>& capture);
-		FLAME_UNIVERSE_EXPORTS void* add_mouse_listener(void (*listener)(void* c, KeyState action, MouseKey key, const Vec2f& pos), const Mail<>& capture);
-		FLAME_UNIVERSE_EXPORTS void* add_drag_and_drop_listener(void (*listener)(void* c, DragAndDrop action, cEventReceiver* er, const Vec2f& pos), const Mail<>& capture);
+		FLAME_UNIVERSE_EXPORTS void* add_mouse_listener(void (*listener)(void* c, KeyState action, MouseKey key, const Vec2i& pos), const Mail<>& capture);
+		FLAME_UNIVERSE_EXPORTS void* add_drag_and_drop_listener(void (*listener)(void* c, DragAndDrop action, cEventReceiver* er, const Vec2i& pos), const Mail<>& capture);
 		FLAME_UNIVERSE_EXPORTS void* add_state_changed_listener(void (*listener)(void* c, EventReceiverState prev_state, EventReceiverState curr_state), const Mail<>& capture);
 
 		FLAME_UNIVERSE_EXPORTS void remove_focus_listener(void* ret_by_add);
@@ -43,8 +43,8 @@ namespace flame
 
 		FLAME_UNIVERSE_EXPORTS void on_focus(FocusType type);
 		FLAME_UNIVERSE_EXPORTS void on_key(KeyState action, uint value);
-		FLAME_UNIVERSE_EXPORTS void on_mouse(KeyState action, MouseKey key, const Vec2f& value);
-		FLAME_UNIVERSE_EXPORTS void on_drag_and_drop(DragAndDrop action, cEventReceiver* er, const Vec2f& pos);
+		FLAME_UNIVERSE_EXPORTS void on_mouse(KeyState action, MouseKey key, const Vec2i& value);
+		FLAME_UNIVERSE_EXPORTS void on_drag_and_drop(DragAndDrop action, cEventReceiver* er, const Vec2i& pos);
 		FLAME_UNIVERSE_EXPORTS void on_state_changed(EventReceiverState prev_state, EventReceiverState curr_state);
 
 		FLAME_UNIVERSE_EXPORTS virtual void start() override;
