+++ source/universe/components/scrollbar.cpp
@@ -88,7 +88,7 @@ namespace flame
 			target_layout = (cLayout*)(parent->parent()->child(0)->find_component(cH("Layout")));
 			assert(target_layout);
 
-			mouse_listener = event_receiver->add_mouse_listener([](void* c, KeyState action, MouseKey key, const Vec2f& pos) {
+			mouse_listener = event_receiver->add_mouse_listener([](void* c, KeyState action, MouseKey key, const Vec2i& pos) {
 				auto thiz = (*(cScrollbarThumbPrivate**)c);
 				if (thiz->event_receiver->active && is_mouse_move(action, key))
 				{
@@ -225,7 +225,7 @@ namespace flame
 
 			auto c_event_receiver = cEventReceiver::create();
 			c_event_receiver->penetrable = true;
-			c_event_receiver->add_mouse_listener([](void* c, KeyState action, MouseKey key, const Vec2f& pos) {
+			c_event_receiver->add_mouse_listener([](void* c, KeyState action, MouseKey key, const Vec2i& pos) {
 				auto thumb = (*(cScrollbarThumb**)c);
 				if (is_mouse_scroll(action, key))
 					thumb->v -= pos.x() * 20.f;
